async_fastapi/pyproject.toml
[tool.poetry]
name = "async-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

# black configuration
# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file
[tool.black]
line_length = 79
skip_string_normalization = true
target_version = ["py311"]

[tool.isort]
profile = "attrs"
combine_as_imports = true
force_alphabetical_sort_within_sections = true
lines_between_types = 0

[tool.poetry.dependencies]
python = "^3.11.0"
fastapi = "0.95.1"
uvicorn = "^0.20.0"
alembic = "^1.8.1"
aiohttp = "^3.8.3"
pytest = "^7.2.0"
websockets = "11.0.3"
greenlet = "^2.0.1"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^1.10.4"}
asyncpg = "^0.27.0"
httpx = "^0.23.3"
pytest-asyncio = "^0.20.3"
black = "^22.12.0"
isort = "^5.11.4"
cryptography = "^39.0.0"
celery = "^5.2.7"
direnv = "^2020.12.3"
psycopg = "^3.1.17"
sqlalchemy-utils = "^0.41.1"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
